Separate App logic to Items component
